chore(test-utils): moves shared utils out of test-file
This should enable better sharing but also limit the tests run (if the
import of test_runem also caused test_runem's test to be run once for
each import, which IIRC used to happen).
